feature: adjusted accessibility, added strings, ajusted bottom sheet components
