+++ components/src/main/java/com/mercadopago/sdk/android/components/bottomsheet/MPBottomSheet.kt
@@ -1,4 +1,4 @@
-package com.mercadopago.sdk.android.components
+package com.mercadopago.sdk.android.components.bottomsheet
 
 import androidx.compose.foundation.background
 import androidx.compose.foundation.clickable
@@ -6,14 +6,14 @@ import androidx.compose.foundation.interaction.MutableInteractionSource
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
 import androidx.compose.foundation.layout.Row
-import androidx.compose.foundation.layout.fillMaxHeight
 import androidx.compose.foundation.layout.fillMaxWidth
 import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.navigationBarsPadding
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.layout.size
 import androidx.compose.foundation.layout.width
-import androidx.compose.foundation.selection.selectableGroup
 import androidx.compose.foundation.shape.RoundedCornerShape
 import androidx.compose.material3.Icon
 import androidx.compose.runtime.Composable
@@ -27,43 +27,36 @@ import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.semantics.Role
 import androidx.compose.ui.semantics.clearAndSetSemantics
 import androidx.compose.ui.semantics.role
-import androidx.compose.ui.semantics.selected
 import androidx.compose.ui.semantics.semantics
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
-import com.mercadopago.sdk.android.components.model.MPBottomSheetListItem
+import com.mercadopago.sdk.android.components.MPText
+import com.mercadopago.sdk.android.components.R
 import com.mercadopago.sdk.android.foundation.theme.MercadoPagoTheme
 import com.mercadopago.sdk.android.foundation.theme.MercadoPagoThemes
 
-private const val MP_BOTTOM_SHEET_GROUP = "MPBottomSheet"
+internal const val MP_BOTTOM_SHEET_GROUP = "MPBottomSheet"
 
 /**
- * Bottom sheet for selecting the cardholder's document type.
- *
- * Renders a drag indicator, a header with [title] and dismiss button, and a list of
- * selectable [items]. The selected item is highlighted with a quiet-fill background
- * and a loud-fill vertical pipe on its left edge.
+ * Generic bottom sheet container with a drag indicator and a header.
  *
  * Intended to be hosted inside a system ModalBottomSheet or equivalent.
+ * Use [MPListBottomSheet] for the document-type selection use case.
  *
  * Note: The design specifies a top-directed shadow (0 -2dp 4dp rgba(40,40,52,0.1)).
  * No foundation token covers this effect — elevation should be handled by the host container.
  *
  * @param title Header title text
- * @param items Ordered list of document type options
- * @param selectedLabel [MPBottomSheetListItem.label] of the currently selected item, or null
- * @param onItemSelected Callback invoked when the user selects an item
  * @param onDismiss Callback invoked when the user taps the dismiss (X) button
  * @param modifier Component modifier
+ * @param content Slot rendered below the header
  */
 @Composable
 fun MPBottomSheet(
     title: String,
-    items: List<MPBottomSheetListItem>,
-    selectedLabel: String?,
-    onItemSelected: (MPBottomSheetListItem) -> Unit,
     onDismiss: () -> Unit,
     modifier: Modifier = Modifier,
+    content: @Composable ColumnScope.() -> Unit,
 ) {
     val topShape = RoundedCornerShape(
         topStart = MercadoPagoTheme.radius.xlarge,
@@ -73,15 +66,12 @@ fun MPBottomSheet(
         modifier = modifier
             .fillMaxWidth()
             .clip(topShape)
-            .background(MercadoPagoTheme.color.surface.primaryIdle),
+            .background(MercadoPagoTheme.color.surface.primaryIdle)
+            .navigationBarsPadding(),
     ) {
         MPBottomSheetDragIndicator()
         MPBottomSheetHeader(title = title, onDismiss = onDismiss)
-        MPBottomSheetList(
-            items = items,
-            selectedLabel = selectedLabel,
-            onItemSelected = onItemSelected,
-        )
+        content()
     }
 }
 
@@ -156,130 +146,26 @@ private fun MPBottomSheetHeader(
     }
 }
 
+@Preview(name = "MPBottomSheet - Custom Content", group = MP_BOTTOM_SHEET_GROUP)
 @Composable
-private fun MPBottomSheetList(
-    items: List<MPBottomSheetListItem>,
-    selectedLabel: String?,
-    onItemSelected: (MPBottomSheetListItem) -> Unit,
-) {
-    Column(
-        modifier = Modifier
-            .fillMaxWidth()
-            .selectableGroup()
-            .padding(MercadoPagoTheme.spacing.paddings.xnano),
-        verticalArrangement = Arrangement.spacedBy(MercadoPagoTheme.spacing.gap.xmicro),
-    ) {
-        items.forEach { item ->
-            MPBottomSheetRow(
-                item = item,
-                selected = item.label == selectedLabel,
-                onItemSelected = onItemSelected,
-            )
-        }
-    }
-}
-
-@Composable
-private fun MPBottomSheetRow(
-    item: MPBottomSheetListItem,
-    selected: Boolean,
-    onItemSelected: (MPBottomSheetListItem) -> Unit,
-) {
-    Box(
-        modifier = Modifier
-            .fillMaxWidth()
-            .clip(MercadoPagoTheme.shape.small)
-            .background(
-                if (selected) {
-                    MercadoPagoTheme.color.interactive.fillQuiet.idle
-                } else {
-                    MercadoPagoTheme.color.interactive.fillMute.idle
-                },
-            )
-            .semantics {
-                role = Role.RadioButton
-                this.selected = selected
-            }
-            .clickable { onItemSelected(item) },
-    ) {
-        MPText(
-            text = item.label,
-            style = MercadoPagoTheme.typography.body.default.large,
-            color = MercadoPagoTheme.color.text.primary,
+internal fun MPBottomSheetPreview() {
+    MercadoPagoTheme(theme = MercadoPagoThemes.Default) {
+        Box(
             modifier = Modifier
-                .fillMaxWidth()
-                .padding(
-                    start = MercadoPagoTheme.spacing.paddings.micro,
-                    top = MercadoPagoTheme.spacing.paddings.xtiny,
-                    bottom = MercadoPagoTheme.spacing.paddings.xtiny,
-                ),
-        )
-        if (selected) {
-            Box(
-                modifier = Modifier
-                    .matchParentSize()
-                    .padding(
-                        start = MercadoPagoTheme.spacing.paddings.xnano,
-                        top = MercadoPagoTheme.spacing.paddings.nano,
-                        bottom = MercadoPagoTheme.spacing.paddings.nano,
-                    ),
+                .background(Color.White)
+                .padding(16.dp),
+        ) {
+            MPBottomSheet(
+                title = "Título do bottom sheet",
+                onDismiss = {},
             ) {
-                Box(
-                    modifier = Modifier
-                        .width(MercadoPagoTheme.borderWidth.medium)
-                        .fillMaxHeight()
-                        .clip(MercadoPagoTheme.shape.full)
-                        .background(MercadoPagoTheme.color.interactive.fillLoud.idle),
+                MPText(
+                    text = "Conteúdo customizado",
+                    style = MercadoPagoTheme.typography.body.default.large,
+                    color = MercadoPagoTheme.color.text.primary,
+                    modifier = Modifier.padding(MercadoPagoTheme.spacing.paddings.xnano),
                 )
             }
-        }
-    }
-}
-
-@Preview(name = "MPBottomSheet - CPF Selected", group = MP_BOTTOM_SHEET_GROUP)
-@Composable
-internal fun MPBottomSheetCpfSelectedPreview() {
-    MercadoPagoTheme(theme = MercadoPagoThemes.Default) {
-        Box(
-            modifier = Modifier
-                .background(Color.White)
-                .padding(16.dp),
-        ) {
-            MPBottomSheet(
-                title = "Documento do titular",
-                items = listOf(
-                    MPBottomSheetListItem(label = "CPF"),
-                    MPBottomSheetListItem(label = "CNPJ"),
-                    MPBottomSheetListItem(label = "Otro"),
-                ),
-                selectedLabel = "CPF",
-                onItemSelected = {},
-                onDismiss = {},
-            )
-        }
-    }
-}
-
-@Preview(name = "MPBottomSheet - No Selection", group = MP_BOTTOM_SHEET_GROUP)
-@Composable
-internal fun MPBottomSheetNoSelectionPreview() {
-    MercadoPagoTheme(theme = MercadoPagoThemes.Default) {
-        Box(
-            modifier = Modifier
-                .background(Color.White)
-                .padding(16.dp),
-        ) {
-            MPBottomSheet(
-                title = "Documento do titular",
-                items = listOf(
-                    MPBottomSheetListItem(label = "CPF"),
-                    MPBottomSheetListItem(label = "CNPJ"),
-                    MPBottomSheetListItem(label = "Otro"),
-                ),
-                selectedLabel = null,
-                onItemSelected = {},
-                onDismiss = {},
-            )
         }
     }
 }
